Extract OTelDocumentBuilder base class from MetricDocumentBuilder
Extract shared document building logic (resource, scope, attributes,
data stream, any-value serialization) into a reusable OTelDocumentBuilder
base class. This allows logs and traces document builders to reuse the
same field-building logic without duplicating code.

Also consolidates the dropped_attributes_count handling into
buildAttributes, and adds addSpanId/addTraceId helpers for logs/traces.
